requirements.txt
appdirs==1.4.4
asgiref==3.2.7
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.7
django-bootstrap3==14.0.0
django-bootstrap4==2.2.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-jquery==3.1.0
django-on-heroku==1.0.1
django-rest-registration==0.5.6
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
dominate==2.5.1
filelock==3.0.12
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.0
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
numpy==1.19.1
Pillow==7.2.0
pipenv==2020.8.13
psycopg2-binary==2.8.5
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
typing==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
visitor==0.1.3
Werkzeug==1.0.1
whitenoise==5.2.0
